vcd_backened/requirements.txt
asgiref==3.8.1
bidict==0.23.1
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.7
Django==4.1.13
django-cors-headers==4.4.0
django-environ==0.11.2
django-mssql-backend==2.8.1
djangorestframework==3.15.2
filelock==3.12.4
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
gTTS==2.5.3
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonify==0.5
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
mysqlclient==2.2.4
numpy==2.1.0
opencv-python==4.10.0.84
pillow==10.4.0
platformdirs==3.11.0
playsound==1.2.2
PyAudio==0.2.14
pycparser==2.22
pycryptodome==3.20.0
pyodbc==5.1.0
pyserial==3.5
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
razorpay==1.4.2
requests==2.32.3
simple-websocket==1.0.0
sounddevice==0.5.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.24.5
Werkzeug==3.0.3
wsproto==1.2.0
